Divide HaBot classes into packages
