+++ src/main/java/HaBot/Command/Command.java
@@ -1,3 +1,10 @@
+package HaBot.Command;
+
+import HaBot.Exception.HaBotException;
+import HaBot.Storage;
+import HaBot.TaskList;
+import HaBot.Ui.Ui;
+
 public class Command {
     public void execute(TaskList taskList, Ui ui, Storage storage) throws HaBotException {
         // does nothing by default
