+++ src/main/java/HaBot/Command/ToDoCommand.java
@@ -1,22 +1,30 @@
+package HaBot.Command;
+
+import HaBot.Exception.HaBotException;
+import HaBot.Storage;
+import HaBot.TaskList;
+import HaBot.Task.ToDo;
+import HaBot.Ui.Ui;
+
 public class ToDoCommand extends Command {
     private final String description;
 
     /**
      * Constructs an AddToDoCommand with the specified description.
      *
-     * @param description The description of the ToDo task.
+     * @param description The description of the HaBot.Task.ToDo task.
      */
     public ToDoCommand(String description) {
         this.description = description.trim();
         if (this.description.isEmpty()) {
-            throw new HaBotException("The description of a ToDo cannot be empty.");
+            throw new HaBotException("The description of a HaBot.Task.ToDo cannot be empty.");
         }
     }
 
     /**
-     * Executes the command to add a ToDo task to the task list.
+     * Executes the command to add a HaBot.Task.ToDo task to the task list.
      *
-     * @param taskList The TaskList to operate on.
+     * @param taskList The HaBot.TaskList to operate on.
      * @param ui The UI to interact with the user.
      * @throws HaBotException If an error occurs during execution.
      */
