+++ src/main/java/HaBot/Exception/HaBotException.java
@@ -1,3 +1,5 @@
+package HaBot.Exception;
+
 public class HaBotException extends RuntimeException {
     public HaBotException(String message) {
         super(message);
