+++ src/main/java/HaBot/Task/ToDo.java
@@ -1,10 +1,12 @@
+package HaBot.Task;
+
 /**
- * Represents a to-do task that has no functional difference from the Task Class.
+ * Represents a to-do task that has no functional difference from the HaBot.Task.Task Class.
  */
 public class ToDo extends Task {
 
     /**
-     * Constructs a ToDo task with the specified description.
+     * Constructs a HaBot.Task.ToDo task with the specified description.
      *
      * @param description The description of the to-do task.
      */
@@ -27,7 +29,7 @@ public class ToDo extends Task {
     public static ToDo fromStoreFormat(String text) {
         String[] parts = text.split("\\|", -1);
         if (parts.length < 3) {
-            throw new IllegalArgumentException("Invalid ToDo format: " + text);
+            throw new IllegalArgumentException("Invalid HaBot.Task.ToDo format: " + text);
         }
         boolean isDone = parts[1].equals("X");
         String description = parts[2].replace("\\|", "|");
